refactor: move datetime field to index.ts
